Take 2: Refactor some t5.data functionality into seqio. Eventually this module will be split off into a new library.
PiperOrigin-RevId: 349280153
